requirement.txt
streamlit
pandas
numpy
textblob
requests
sqlalchemy
psycopg2-binary